Move GHA yaml generation script fully into the CLI
