+++ setup.py
@@ -126,7 +126,8 @@ setup(
                             'ygginstall=yggdrasil.command_line:ygginstall',
                             'yggclean=yggdrasil.command_line:yggclean',
                             'yggmodelform=yggdrasil.command_line:yggmodelform',
-                            'yggdevup=yggdrasil.command_line:yggdevup'],
+                            'yggdevup=yggdrasil.command_line:yggdevup',
+                            'ygggha=yggdrasil.command_line:generate_gha_workflow'],
     },
     license="BSD",
     python_requires='>=3.5',
